Fix: add missing quotes for script dir
When `getopts` finds an option starting with "+" it takes the rest of the
option and converts it to a number.  The algorithm to find the index of
the last character is to subtract 1 from the string length, which gives 0
on a lone "+". (string.substring s 1 0) then throws an exception.

When `getopts` is run from the `femto.rc` file throwing this exception
results in a segfault, which is the true bug.

The current commit just handles the special case of a lone "+" option in
`getopts`.