Fix blocking shell-command. Make C code less, Lisp code more, Add keybindings.
`shell-command` was not closing stdin, so it blocked when the sub process
read from it.

Instead of fixing this in C code, a new primitive `system` is introduced.
as the building block for arbitrary execution of subprocesses.

The new `shell-command` constructs the commandline from Lisp and simply
adds " <&-" at the end which closes stdin.

Instead of implementing a string formatting library it was decided to make
`concat` smarter, so it converts numbers to strings and traverses sublists.

A helper function `stringify` which takes one argument and tries to convert
it to a string is used in `concat`.

Implementation of `prompt-filename` was incidential, but I guess it fit's
well when trying to create a command line.

Since we needed to insert file content into the *output* buffer of
`shell-command`, the C code `insert-file` was also replaced by a new
primitive `insert-file-content-literally` and a Lisp level implementation
of `insert-file`.

----
The new primitives give way for implementing Elisp like `call-process` and
`call-process-region`.