Add batch and debug mode, select init_file, formatting, signal primitive
# Batch Mode

When the environment variable FEMTO_BATCH exists, `femto` runs in batch mode:
- GUI is not started
- All Lisp output is sent to stdout

# Debug Mode

When the environment variable FEMTO_DEBUG exists, `femto` turns on debugging.
Some details of the inner working are logged to the file `debug.out`.

# Select init_file at runtime

When the environment variable FEMTORC is set, `femto` interprets its value as the path
to the init file and tries to load it instead of the default init_file.

# Other Changes

## Formatting

Femto coding style will be K&R with 4 space indents, with exception of `lisp.c` which
remains with tab indents.  `main.c` has been reformatted.

 ## main.c

Has been restructured completely so that GUI relevant functionality can be separated from
other logic.

load_config() does not `fatal`ly exit when it fails, instead an error is shown in the message
area.  Femto can work very well without the initialization file.

debug() got its `vsnprintf` for which it has asked already in a comment.

## lisp.c

A new primitive `(signal 'error-symbol error-details)` has been added.  It throws an exception.
The interface is built analogous to Elisp.  It was just used for testing, but might be useful in
the future anyway.

# Motivation

Batch mode and init file selection will make unit testing easier.

Init file selection (and debug mode) will simplify debugging of startup problems.