Restore shell-command call interface
Originally there was a builtin (shell-command string). Then it was replaced
by the (system string) builtin and an emulation provided in Lisp. However
the emulation did not accept a commandline argument.

The new implementation accepts 0 or more arguments. If called with no
argument it asks for user input and feeds it as a string to
(exec-shell string), otherwise it feeds the argument to exec-shell.