Select script_dir when running femto. Cleanups
SCRIPTDIR needs not be specified with quotes anymore, it is stringified
in main.c as default for `script_dir` and handed over to init_lisp(). If
FEMTOLIB exists its value is taken instead.

femto.rc does not need to be generated anymore out of femto.rc.in, the latter
is removed.

Cleanup:

debug() messages have been adapted to "function(param): " format.

Whitespace cleanup in lisp.c

An extra close() in e_load() has been removed to avoid errors.