Add (require) and (memq), move Lisp examples to subdirectory, re-add femto.rc
Instead of (load-script)ing a library file, it can be (require)'d as a
"feature".
(require) remembers loaded files and does not load them more then once.

Library files must end with (provide 'feature) where feature is the
file name w/o the extension.

Caution: Some of the provided library files segfault when (require)'d
but not when (load-script)ed. This has to be investigated.

(memq) finds an item in a list. It is needed by (require).

`femto.rc` is reordered, so that core functions come first.