Make (cond) compatible with Elisp (and others), update unit test
Tiny Lisp (cond) does not allow a clause to be nil, stating (wrongly) that
nil is not a list.  This causes esceptions to be thrown where not expected.

A new implementation is provided, which tries to be compatible with Elisp.
Additionally we are saving some nested calls, because we look out for
explicit nil in several places and handle them directly, while Tiny Lisp
uses recursion.

The implementations differ in this cases:
(cond nil) ;; nil - TL: error: nil is not a list
(cond ()) ;; nil - TL: error: nil is not a list