Re-enable git and oxo by default, remove trap invocation.
The new exception handling code together with increased memory allows
the editor to start up correctly while loading both libraries.

While the garbage collector might still trigger a segfault, it does not
with the current memory settings 'on my machine' (tm).

The `trap` function did not work and is not needed anymore. It was removed
from the lisp interpreter and from here.