Add stream object type to Lisp interpreter
Prepare a switch to I/O from emulated streams to OS (POSIX) streams.

POSIX streams are already capable to read both from files and from
memory (strings).  This first steps introduces a TYPE_STREAM Lisp
object, which holds and manages POSIX streams.

The core interpreter will not expose these objects on the Lisp level, it
is only used internally and for interfacing with Lisp on C level. Therefore
only the fopen() and fclose() primitives are needed.

An extension will provide both the input and output functions and expose
all of them on the script level.