Add rmkx mode esc sequence for home and end key
Note: in order to make home and end keys to work in rmkx mode
(xfce4-terminal default?)  they had to be defined *first* in the
keybindings, otherwise one of the two keys would not work. Now
one of them does not work in skmx mode.  An obvious reason for
this behavior was not apparent.