Add type-of and move type queries to Lisp code, except consp.
Type as Lisp objects allows for a trivial implementation of type-of.

This can be used to implement all type checkers - including the ones not
yet implemented - in Lisp, they are found at the start of the core.lsp
library.

consp is maintained as primitive for speed, because it is the most used in
the core, specifically in meta programming.