WIP: add double float extension
Double floats are back. Double arithmetic operations are prefixed with 'd':
d+, d- ...

The build system currently only builds with the double extension enabled.

Type coercion has to be implemented in Lisp, to be able to use the "normal"
arithmetic operators with double numbers.

This commit introduces another mayor change to the inner workings of fLisp:

Instead of an integer offset into the primitives array, the pointer to the
primitive is stored in the primitive lisp object.  The redundante `name`
field is removed.

Bug Fixes:
- Reading double number never worked: the decimal point was not read in, so
  doubles would always be "multiplied" to have no decimal fraction on read.
- Removing the name field revealed a bug in the usage of the result value,
  which lead to segmentation faults.