Move eq from C to Lisp, introduce same and string-equal primitive
eq is either object equality, or depending on type a certain type equality.
type equality is baked in in C and therefore redundant.

Strings are string compared, integers and doubles are arithmetically
compared.

`same` is pure object equality, it is always nil on different but "equal"
number and string constants.

String equality did not have it's own test, integers and doubles have.

`string-equal` is a standard primitive in Lisp and added to the core.

`same` and `string-equal` together with number comparison allow to
 implement `eq` in Lisp.